requirements.txt
pybit
pandas
python-dotenv
matplotlib
mplfinance